Move reference search utils to src/searches
